requirements.txt
django
python-binance
django-debug-toolbar
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pillow
six
djangorestframework